+++ internal/core/humanize.go
@@ -1,91 +1,86 @@
-// SPDX-FileCopyrightText: 2005
-// 2008 SAP SE or an SAP affiliate company
+// SPDX-FileCopyrightText: 2026 SAP SE or an SAP affiliate company
 // SPDX-License-Identifier: Apache-2.0
 
 package core
 
 import (
-	"math"
-	"slices"
 	"strconv"
 
 	"github.com/sapcc/go-api-declarations/limes"
 )
 
-// suitableUnit takes a value and its current unit, and returns a suitable
-// human friendly unit that this value can be converted to.
+// ValueFormatter is the interface for a function that renders integer values into their output representation in CSV.
+// This may involve a conversion between units, based on the unit that was chosen for the respective CSV row during humanization.
+type ValueFormatter func(uint64) string
+
+// DefaultValueFormatter is a ValueFormatter that does not perform unit conversion.
+func DefaultValueFormatter(value uint64) string {
+	return strconv.FormatUint(value, 10)
+}
+
+var (
+	allCleanUnitsBasedOnNone = []limes.Unit{
+		limes.UnitNone,
+	}
+	allCleanUnitsBasedOnBytes = []limes.Unit{
+		// This is sorted in descending order of size, so the first viable choice
+		// will be the "best" choice (as in: producing the smallest number values).
+		// e.g. 2 GiB is better than 2048 MiB is better than 2097152 KiB
+		limes.UnitExbibytes,
+		limes.UnitPebibytes,
+		limes.UnitTebibytes,
+		limes.UnitGibibytes,
+		limes.UnitMebibytes,
+		limes.UnitKibibytes,
+		limes.UnitBytes,
+	}
+)
+
+// PickHumanizedValueFormatter takes several values that are to be interpreted in the given unit,
+// and chooses the best unit to format them in for human-readable values.
 //
-// E.g. A value 82854982 with unit limes.UnitBytes will return
-// limes.UnitMebibytes.
-func suitableUnit(v uint64, u limes.Unit) limes.Unit {
-	if v < 10 {
-		return u
+// For example, `PickHumanizedValueFormatter([2048 4096], UnitBytes)` returns
+// UnitKibibytes and a formatter that renders 2048 as 2 and 4096 as 4.
+func PickHumanizedValueFormatter(unit limes.Unit, values []uint64) (limes.Unit, ValueFormatter) {
+	// possible choices of result units
+	baseUnit, multiplierToBase := unit.Base()
+	var possibleUnits []limes.Unit
+	switch baseUnit {
+	case limes.UnitNone:
+		possibleUnits = allCleanUnitsBasedOnNone
+	case limes.UnitBytes:
+		possibleUnits = allCleanUnitsBasedOnBytes
+	default:
+		// if Limes adds new base units that we don't know about, not humanizing is a safe fallback
+		return unit, DefaultValueFormatter
 	}
 
-	var sizes []limes.Unit
-	for i, unit := range LimesUnits {
-		if u == unit {
-			sizes = LimesUnits[i:]
-			break
+	// pick the first candidate unit that produces clean integers for all presented values
+UNIT:
+	for _, targetUnit := range possibleUnits {
+		targetBaseUnit, multiplierToTarget := targetUnit.Base()
+		if targetBaseUnit != baseUnit {
+			// defense in depth: the lists of `possibleUnits` above should be prepared such that all the base units match
+			continue
 		}
-	}
-
-	base := 1024.0
-	logn := math.Log(float64(v)) / math.Log(base)
-	e := math.Floor(logn)
-	return sizes[int(e)]
-}
-
-// convertValue converts a value from its old unit to the new unit.
-// The returned value is rounded off.
-//
-// Use suitableUnit() to find an appropriate new unit for a specific value.
-func convertValue(v uint64, source, target limes.Unit) string {
-	if v == 0 {
-		return "0"
-	}
-	_, sourceMultiple := source.Base()
-	_, targetMultiple := target.Base()
-	valueInBase := v * sourceMultiple
-	newV := float64(valueInBase) / float64(targetMultiple)
-	newV = math.Round(newV*100) / 100 // round to second decimal place
-	return strconv.FormatFloat(newV, 'f', -1, 64)
-}
-
-// valToStrFunc is a function that can be used to convert a resource value to a
-// string.
-type valToStrFunc func(v uint64) string
-
-var defaultValToStrFunc valToStrFunc = func(v uint64) string { return strconv.FormatUint(v, 10) }
-
-// getValToStrFunc finds a human friendly unit for the given vals and returns a
-// function that can be used to convert resource values to this new unit.
-func getValToStrFunc(humanize bool, old limes.Unit, vals []uint64) (f valToStrFunc, newUnit limes.Unit) {
-	f = defaultValToStrFunc
-	newUnit = old
-	if humanize && old != limes.UnitNone {
-		// Find a new human friendly unit based on the smallest value.
-		if sml := smallestValue(vals); sml > 0 {
-			if newUnit = suitableUnit(sml, old); newUnit != old {
-				f = func(v uint64) string {
-					return convertValue(v, old, newUnit)
-				}
+		for _, value := range values {
+			rawValue := value * multiplierToBase
+			if rawValue/multiplierToBase != value {
+				// if conversion to the base unit runs into an integer overflow, not humanizing is a safe fallback
+				return unit, DefaultValueFormatter
+			}
+			// the value in the target unit would be `rawValue / multiplierToTarget`, does this work within uint64?
+			if rawValue%multiplierToTarget != 0 {
+				continue UNIT
 			}
 		}
-	}
-	return f, newUnit
-}
-
-func smallestValue(vals []uint64) uint64 {
-	nonZero := make([]uint64, 0, len(vals))
-	for _, v := range vals {
-		if v > 0 {
-			nonZero = append(nonZero, v)
+		// all `values` could be converted cleanly into `targetUnit`
+		return targetUnit, func(value uint64) string {
+			convertedValue := (value * multiplierToBase) / multiplierToTarget
+			return strconv.FormatUint(convertedValue, 10)
 		}
 	}
-	if len(nonZero) == 0 {
-		return 0
-	}
-	slices.Sort(nonZero)
-	return nonZero[0]
+
+	// defense in depth: if (somehow!) no candidate was viable, not humanizing is a safe fallback
+	return unit, DefaultValueFormatter
 }