+++ internal/core/project_rates.go
@@ -90,10 +90,10 @@ func (p ProjectRatesReport) render(opts *OutputOpts) CSVRecords {
 				defaultWindow = pSrvRate.DefaultWindow.String()
 			}
 
-			valToStr := defaultValToStrFunc
+			formatter := DefaultValueFormatter
 			if opts.CSVRecFmt == CSVRecordFormatLong {
 				r = append(r, p.DomainID, p.DomainName, p.UUID, p.Name, pSrv.Area, string(pSrv.Type), string(pSrvRate.Name),
-					valToStr(pSrvRate.Limit), valToStr(pSrvRate.DefaultLimit), window, defaultWindow,
+					formatter(pSrvRate.Limit), formatter(pSrvRate.DefaultLimit), window, defaultWindow,
 					pSrvRate.UsageAsBigint, pSrvRate.Unit.String(), timestampToString(pSrv.ScrapedAt),
 				)
 			} else {
@@ -104,7 +104,7 @@ func (p ProjectRatesReport) render(opts *OutputOpts) CSVRecords {
 					domainNameOrID = p.DomainName
 				}
 				r = append(r, domainNameOrID, projectNameOrID, string(pSrv.Type), string(pSrvRate.Name),
-					valToStr(pSrvRate.Limit), window, pSrvRate.UsageAsBigint, pSrvRate.Unit.String(),
+					formatter(pSrvRate.Limit), window, pSrvRate.UsageAsBigint, pSrvRate.Unit.String(),
 				)
 			}