+++ CHANGELOG.md
@@ -13,11 +13,17 @@ The sections should follow the order `Added`, `Changed`, `Fixed`, `Removed`, and
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.7.0...HEAD)
+## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.11.0...HEAD)
+
+### Added
 
 ### Changed
 
-## 3.11.0 - 2026-03-27
+## 3.11.0 - 2026-03-30
+
+### Added
+
+- Added support for resources defined in terms of non-standard units (e.g. "128 GiB" instead of just "GiB" or "TiB").
 
 ### Changed